% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adviceDownload.R
\name{adviceDownload}
\alias{adviceDownload}
\title{Download advice scenarios}
\usage{
adviceDownload(adviceKeys, dir = ".")
}
\arguments{
\item{adviceKeys}{a vector of advice view ids.}

\item{dir}{directory to download to.}
}
\description{
Download a zip file of a set of advice scenarios
}
\examples{
\dontrun{
adviceDownload(c(1009, 1011, 1022))
}

}
