% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpUtils.R
\name{api}
\alias{api}
\title{Create The ASD Webservice URL}
\usage{
api(
  stock_name = NULL,
  year = NULL,
  adviceKey = NULL,
  api = c("record", "table", "notes")
)
}
\arguments{
\item{stock_name}{string value of the stock code you wish to look}

\item{year}{integer value of assessment year}

\item{adviceKey}{integer value of stock adviceKey}

\item{api}{string value of different entry types to choose from ("record", "table", "notes")}
}
\value{
string containing URL
}
\description{
Returns a string with the webservice URL to be used depending on the ASD entry type.
}
\examples{
\dontrun{
api(stock = "cod.27.47d20", year = 2022, api = "record")
api(adviceKey = 2837, api = "table")
api(adviceKey = 2837, api = "notes")
}

}
\references{
https://sg.ices.dk/adviceview/AdviceList
}
