% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asd_api.R
\name{asd_api}
\alias{asd_api}
\title{Create The ASD Webservice URL}
\usage{
asd_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{named arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

asd_api("hi", bye = 21)
asd_api("csquares", argument1 = 2021, argument2 = NULL)


}
