% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNotes.R
\name{checkNotes}
\alias{checkNotes}
\title{Check catch scenario notes in a browser}
\usage{
checkNotes(notes)
}
\arguments{
\item{notes}{df of culumns element of catch scenario table}
}
\description{
Check catch scenario notes in a browser
}
\examples{
\dontrun{
checkNotes(notes)
}

}
