% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTable.R
\name{checkTable}
\alias{checkTable}
\title{Check catch scenario table in a browser}
\usage{
checkTable(columns, rows, values)
}
\arguments{
\item{columns}{df of culumns element of catch scenario table}

\item{rows}{df of rows element of catch scenario table}

\item{values}{df of values element of catch scenario table}
}
\description{
Check catch scenario table in a browser
}
\examples{
data(scenarioColumns)
data(scenarioRows)
data(scenarioValues)
checkTable(scenarioColumns, scenarioRows, scenarioValues)

}
