% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createASDxml.R
\name{createASDxml}
\alias{createASDxml}
\title{Create the XML format for catch scenario data}
\usage{
createASDxml(
  info,
  scenarioColumns,
  scenarioRows,
  scenarioValues,
  scenarioNotes
)
}
\arguments{
\item{info}{details}

\item{scenarioColumns}{details}

\item{scenarioRows}{details}

\item{scenarioValues}{details}

\item{scenarioNotes}{details}
}
\description{
Create the XML format for catch scenario data
}
\examples{
\dontrun{
checkTable(col, rows, vals)
}

}
