% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_advice_view_info.R
\name{get_advice_view_info}
\alias{get_advice_view_info}
\title{Get Advice Record}
\usage{
get_advice_view_info(stock_name, year)
}
\arguments{
\item{stock_name}{string value of the stock code you wish to look}

\item{year}{integer value of assessment year}
}
\value{
data.frame containing metadata for the advice view record
}
\description{
Returns a df with the metadata for the advice view record for a specific stock and year. It uses the advice view web-service.
}
\examples{
\dontrun{
get_advice_view_info("bll.27.3a47de", 2021)
}

}
\references{
https://sg.ices.dk/adviceview/AdviceList
}
