% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch_scenario_notes.R
\name{get_catch_scenario_notes}
\alias{get_catch_scenario_notes}
\title{Get Advice Catch Scenario Table Footnotes}
\usage{
get_catch_scenario_notes(adviceKey)
}
\arguments{
\item{adviceKey}{integer value of stock adviceKey}
}
\value{
data.frame containing the catch scenarios table footnotes and their symbols
}
\description{
Returns a data.frame containing the catch scenario table's footnotes. It uses an ASD dedicated web-service.
}
\examples{
\donttest{
get_catch_scenario_notes(3056)
}

}
\references{
https://sg.ices.dk/adviceview/AdviceList
}
