% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch_scenario_table.R
\name{get_catch_scenario_table}
\alias{get_catch_scenario_table}
\title{Get Advice Catch Scenario Table}
\usage{
get_catch_scenario_table(adviceKey, assessmentYear)
}
\arguments{
\item{adviceKey}{integer value of stock adviceKey}

\item{assessmentYear}{integer value of assessment year}
}
\value{
data.frame containing the catch scenarios table for the advice view record
}
\description{
Returns the catch scenario table for the selected stock and year using the advice view web service.
}
\examples{
\donttest{
get_catch_scenario_table(3056,2022)
}

}
\references{
https://sg.ices.dk/adviceview/AdviceList
}
