% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASD.R
\name{getASD}
\alias{getASD}
\title{Get Any ASD Data}
\usage{
getASD(
  stock = NULL,
  year = NULL,
  assessmentKey = NULL,
  outputData = "table",
  purpose = "Advice"
)
}
\arguments{
\item{stock}{a stock name, e.g. cod-347d, or cod to find all cod stocks, or
NULL to process all stocks.}

\item{year}{the assessment year, e.g. 2015, or 0 to process all years.}

\item{assessmentKey}{the assessmentKey of the stock of interest.}

\item{outputData}{the data you want to retrieve, either "table", "record", or "notes".}

\item{purpose}{the purpose of the entry, options are "Advice", "Bench",
"InitAdvice", default is "Advice".}
}
\value{
A data frame (default) or a list if \code{combine} is \code{TRUE}.
}
\description{
This function combines the functionality of \code{\link[icesSAG]{getListStocks}},
\code{\link[icesSAG]{getFishStockReferencePoints}}, \code{\link[icesSAG]{getSummaryTable}}
and \code{\link[icesSAG]{getStockDownloadData}}.
It supports querying many stocks and years in one function call.
}
\note{
Only years with "Published" status are returned.
}
\examples{
\dontrun{
# Example usage
getASD(stock = "cod.27.7e-k", year = 2023, outputData = "record")
# Example usage with assessmentKey
getASD(assessmentKey = 18719, outputData = "table")
}
}
\seealso{
\code{\link[icesSAG]{getListStocks}}

\code{\link[icesSAG]{findAssessmentKey}} finds lookup keys.

\code{\link[icesSAG]{icesSAG-package}} gives an overview of the package.
}
\author{
Luca Lamoni.
}
