% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchScenariosNotes.R
\name{getCatchScenariosNotes}
\alias{getCatchScenariosNotes}
\title{Get Catch Scenarios Notes}
\usage{
getCatchScenariosNotes(adviceKey)
}
\arguments{
\item{adviceKey}{advice view id}
}
\description{
Get Catch Scenarios Notes for a given record
}
\examples{
\rundonttest{
catch_scenario_notes <- getCatchScenariosNotes(2814)
}

}
