% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesASD-package.R
\name{icesASD-package}
\alias{icesASD-package}
\alias{icesASD}
\title{Stock Advice and Scenarios Database Web Services}
\description{
R interface to access the web services of the ICES Stock Advice
and Scenarios database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getAdviceViewRecord}}             \tab Get Advice Record\cr
  \code{\link{getCatchScenariosTable}}             \tab Get Advice Catch Scenario Table\cr
 \code{\link{getCatchScenariosNotes}}             \tab Get Advice Catch Scenario Table Footnotes

}
}
\references{
ICES Advice and Scenarios database: \url{http://sg.ices.dk}.

ICES Advice and Scenarios database web services:
\url{http://sg.ices.dk/webservices.aspx}.
}
\author{
Colin Millar, Luca Lamoni, Carlos Pinto, Laura Petre, David Miller and Neil Maginis.
}
