% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesSAS-package.R
\docType{package}
\name{icesSAS-package}
\alias{icesSAS-package}
\alias{icesSAS}
\title{Stock Advice and Scenarios Database Web Services}
\description{
R interface to access the web services of the ICES Stock Advice
and Scenarios database.
}
\details{
\emph{Get dataset:}
\tabular{ll}{
  \code{\link{getSummaryTable}}             \tab summary results\cr
  \code{\link{getFishStockReferencePoints}} \tab reference points\cr
  \code{\link{getSAG}}                      \tab any data
}
\emph{Look up codes:}
\tabular{ll}{
  \code{\link{findAssessmentKey}}       \tab find assessment key\cr
  \code{\link{getListStocks}} \tab list of stocks
}
}
\references{
ICES Stock Advice and Scenarios database: \url{http://sg.ices.dk}.

ICES Stock Advice and Scenarios database web services:
\url{http://sg.ices.dk/webservices.aspx}.
}
\author{
Colin Millar, Luca Lamoni, Carlos Pinto, Laura Petre, David Miller and Neil Maginis.
}
