% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAsdUpload.R
\name{readAsdUpload}
\alias{readAsdUpload}
\title{Read a file in the ASD upload format}
\usage{
readAsdUpload(filename, format = FALSE)
}
\arguments{
\item{filename}{the file path of the file to read}

\item{format}{should the output be formatted into an easy to read table}
}
\description{
Read a file in the ASD upload format, returning a list of metadata and data.frame
}
\examples{
filename <- system.file("CatchScenarios_AdviceKey_3207.xml", package = "icesASD")
asd <- readAsdUpload(filename)


}
