% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sets.R
\docType{data}
\name{scenarioValues}
\alias{scenarioValues}
\title{Catch Scenario Datasets}
\format{
List containing four entries:
\tabular{ll}{
  \code{stockCode}  \tab ICES fish stock code\cr
  \code{AssessmentYear}   \tab assessment year\cr
  \code{Purpose} \tab assessment purpose\cr
  \code{contact} \tab contact person
}
}
\source{
ICES (2024).
}
\usage{
scenarioValues
}
\description{
Predefined colors that can be useful in TAF plots.
}
\details{
The data are an excerpt from ...
}
