% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdviceViewRecord.R
\name{getAdviceViewRecord}
\alias{getAdviceViewRecord}
\title{Get Advice View Record}
\usage{
getAdviceViewRecord()
}
\description{
Get Advice View Record
}
\examples{
\rundonttest{
catch_scenario_list <- getAdviceViewRecord()
}
}
