% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdviceViewRecord.R
\name{getAdviceViewRecord}
\alias{getAdviceViewRecord}
\title{Get Advice View Record}
\usage{
getAdviceViewRecord(stockcode = NULL, year = NULL)
}
\arguments{
\item{stockcode}{string value of the stock code you wish to look}

\item{year}{integer value of assessment year}
}
\description{
Get Advice View Record
}
\examples{
\rundonttest{
catch_scenario_list <- getAdviceViewRecord()
head(catch_scenario_list, 3)

getAdviceViewRecord("bll.27.3a47de", 2021)
}

}
