% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchScenariosNotes.R
\name{getCatchScenariosNotes}
\alias{getCatchScenariosNotes}
\title{Get Catch Scenarios Notes}
\usage{
getCatchScenariosNotes(adviceKey)
}
\arguments{
\item{adviceKey}{advice view id(s)}
}
\description{
Get Catch Scenarios Notes for a given record
}
\examples{
\donttest{
catch_scenario_note <- getCatchScenariosNotes(2814)
catch_scenario_notes <- getCatchScenariosNotes(c(2814, 2815))
}

}
