% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCatchScenariosTable.R
\name{getCatchScenariosTable}
\alias{getCatchScenariosTable}
\title{Get Catch Scenarios Table}
\usage{
getCatchScenariosTable(adviceKey)
}
\arguments{
\item{adviceKey}{advice view id}
}
\description{
Get Catch Scenarios Table for a given record
}
\examples{
\donttest{
catch_scenario_table <- getCatchScenariosTable(2814)
}

}
